move code samples to reusable components
